Зайцы почти готовы. Волки начали ходить
Зайцы ходят и плодятся. Осталась шлифовка и оптимизация. Волки пока ходят по пустым клеткам, но ходят

+++ settings.h
@@ -6,7 +6,7 @@
 
 struct Settings {
 	int map_length;
-	float reprod_chance;
+	int reprod_chance;
 	int wolf_health;
 	int rab1, rab2;
 	int wlf_m1, wlf_m2;
